[#34] refactor: DTO(User), DB문자셋 옵션 추가
1. User DTO Optional Chaining 적용
- DAO에서 객체 반환 때, 결과값이 없어도 클래스 생성이 되도록 함.

2. 기타 수정
- DAO 파일 참조 수정
- DB 문자셋 지정
