+++ dao/dao.js
@@ -1,4 +1,4 @@
 const DataAccessObject = require('./DataAccessObject');
-const pool = require('./poolOptions').production;
+const pool = require('./constants/poolOptions').production;
 
 module.exports = new DataAccessObject(pool);
