+++ dao/dto/User.js
@@ -1,8 +1,8 @@
 class User {
   constructor(attrs) {
-    this.id = attrs.id;
-    this.name = attrs.name;
-    this.profile_image = attrs.profile_image;
+    this.id = attrs?.id;
+    this.name = attrs?.name;
+    this.profile_image = attrs?.profile_image;
   }
 }
